[GH] [CODE] Save system, Game logic

+++ Buissnes/Assets/Scripts/Runtime/Save/ConcreteModel/Player/ConcretePlayerModel.cs
@@ -3,6 +3,6 @@
     [System.Serializable]
     public class ConcretePlayerModel
     {
-        public int Balance;
+        public float Balance;
     }
 }
